Exploration of mechanisms used for resource protection.
In deadlock.cpp we can see how and what causes deadlock situation, how to prevent it and some good practices. Unique lock is used for fine grain resource protection. Condition variable is used when we are waiting for other thread/s to initialize variables we need. Future is mechanism used for trans-thread resources not ready and setted up yet.